media-gfx/utsushi: ImageMagick deskew/crop support
Fixes to allow utsushi to be built with --magick-pp.  According to:
  https://github.com/utsushi/utsushi/issues/43

The API changes from ImageMagick 6 to 7 means build failures as there
are old API's in use.  It seems, that these API's are not actually used
(as in GraphicsMagick doesn't use them), so it may be ok to disable the
checks for them in 'doc-locate.cpp'.

This is not yet confirmed, but allows building and using deskew in my
own testing.

Additionally, minor fixes/changes to 'configure' flags.